requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database Drivers
neo4j==5.13.0
asyncpg==0.29.0
motor==3.3.2
pymilvus==2.3.4

# Database ORM
sqlalchemy==2.0.23

# Embeddings and ML
sentence-transformers==2.2.2
numpy==1.24.4

# LLM APIs
anthropic==0.7.1
openai==1.3.0

# Security & Authentication
PyJWT==2.11.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Async
aiofiles==23.2.1

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
